+++ package.json
@@ -17,7 +17,7 @@
     "packages/generator",
     "packages/test-util",
     "packages/util",
-    "packages/rest-generator",
+    "packages/openapi-generator",
     "test-packages/integration-tests",
     "test-packages/e2e-tests",
     "test-packages/test-services",
@@ -53,7 +53,7 @@
     "analytics": "yarn workspace @sap-cloud-sdk/analytics",
     "core": "yarn workspace @sap-cloud-sdk/core",
     "generator": "yarn workspace @sap-cloud-sdk/generator",
-    "rest-generator": "yarn workspace @sap-cloud-sdk/rest-generator",
+    "openapi-generator": "yarn workspace @sap-cloud-sdk/openapi-generator",
     "test-util": "yarn workspace @sap-cloud-sdk/test-util",
     "util": "yarn workspace @sap-cloud-sdk/util",
     "e2e-tests": "yarn workspace @sap-cloud-sdk/e2e-tests",
