+++ packages/core/src/index.ts
@@ -4,4 +4,4 @@ export * from './http-client';
 export * from './odata-common';
 export * from './odata-v2';
 export * from './odata-v4';
-export * from './rest';
+export * from './openapi';
