+++ packages/core/src/openapi/types.ts
