--- packages/core/src/rest/index.ts
@@ -1,1 +1,0 @@
-export * from './rest-request-builder';
