+++ packages/openapi-generator/bin/run
