+++ packages/openapi-generator/bin/run.cmd
