+++ packages/openapi-generator/src/generator.spec.ts
