+++ packages/openapi-generator/src/openapi-types.ts
