+++ packages/openapi-generator/src/parser/document.spec.ts
