+++ packages/openapi-generator/src/parser/document.ts
