+++ packages/openapi-generator/src/parser/operation.spec.ts
