+++ packages/openapi-generator/src/parser/operation.ts
