+++ packages/openapi-generator/src/parser/refs.spec.ts
