+++ packages/openapi-generator/src/parser/refs.ts
