+++ packages/openapi-generator/src/templates/baseApi.mustache
