+++ packages/openapi-generator/src/templates/configuration.mustache
