+++ packages/openapi-generator/src/wrapper-files/api-file.spec.ts
