+++ packages/openapi-generator/src/wrapper-files/create-file.spec.ts
