+++ packages/openapi-generator/src/wrapper-files/index-file.spec.ts
