+++ packages/openapi-generator/src/wrapper-files/index.ts
