+++ scripts/generate-test-services.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import util from 'util';
 import { generate as generateOdata } from '../packages/generator/src';
-import { generateClients as generateOpenApi } from '../packages/rest-generator/src';
+import { generateClients as generateOpenApi } from '../packages/openapi-generator/src';
 import { ODataVersion } from '../packages/util/src';
 
 type fsTypes = typeof fs.readdir & typeof fs.writeFile & typeof fs.readFile;
