+++ test-packages/test-services/rest/test-service/api.ts
@@ -3,28 +3,28 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import { RestRequestBuilder } from '@sap-cloud-sdk/core';
+import { OpenApiRequestBuilder } from '@sap-cloud-sdk/core';
 import { DefaultApi } from './openapi/api';
 import { TestEntity } from './openapi/model';
 
 export const TestServiceApi = {
-  getAllEntities: (stringParameter?: string, integerParameter?: number) => new RestRequestBuilder<DefaultApi, 'getAllEntities'>(
+  getAllEntities: (stringParameter?: string, integerParameter?: number) => new OpenApiRequestBuilder<DefaultApi, 'getAllEntities'>(
     DefaultApi,
     'getAllEntities',
     stringParameter,
     integerParameter
   ),
-  createEntity: (testEntity?: TestEntity) => new RestRequestBuilder<DefaultApi, 'createEntity'>(
+  createEntity: (testEntity?: TestEntity) => new OpenApiRequestBuilder<DefaultApi, 'createEntity'>(
     DefaultApi,
     'createEntity',
     testEntity
   ),
-  getEntityByKey: (entityId: string) => new RestRequestBuilder<DefaultApi, 'getEntityByKey'>(
+  getEntityByKey: (entityId: string) => new OpenApiRequestBuilder<DefaultApi, 'getEntityByKey'>(
     DefaultApi,
     'getEntityByKey',
     entityId
   ),
-  countEntities: () => new RestRequestBuilder<DefaultApi, 'countEntities'>(
+  countEntities: () => new OpenApiRequestBuilder<DefaultApi, 'countEntities'>(
     DefaultApi,
     'countEntities'
   )
